requirements.txt
# requirements.txt

# Librería para el parseo avanzado de XML en la Fase 2
lxml==6.0.0
google-generativeai
python-dotenv
markdown2

# RAG: Vector DB, Embeddings y Carga de Documentos
chromadb
sentence-transformers
langchain
pypdf
PyMuPDF